+++ CHANGELOG.md
@@ -16,7 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0]
 
 ### Added
-- Added cowvr strategy
 - Added new default footprint algorithm
 ### Deprecated
 ### Removed
@@ -27,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Initial implementation for forge-py to footprint SCATSAT1_ESDR_L2_WIND_STRESS_V1.1
+- Added cowvr strategy
 ### Deprecated
 ### Removed
 ### Fixed