+++ CHANGELOG.md
@@ -13,10 +13,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 
+## [0.2.0]
+
+### Added
+- Update default footprint algorithm, now there is only one algorithm for forge-py
+- Poetry update python libraries
+### Deprecated
+### Removed
+### Fixed
+
+
 ## [0.1.0]
 
 ### Added
 - Initial implementation for forge-py to footprint SCATSAT1_ESDR_L2_WIND_STRESS_V1.1
+- Added cowvr strategy
 ### Deprecated
 ### Removed
 ### Fixed